# [Kernel] Update Cutlass int8 kernel configs for SM80