--- a/csrc/quantization/cutlass_w8a8/scaled_mm_c2x.cu
+++ b/csrc/quantization/cutlass_w8a8/scaled_mm_c2x.cu
@@ -300,56 +300,58 @@ struct sm80_config<InType, OutType, Epilogue, 16> {
 
 }  // namespace
 
-template <typename InType, typename OutType>
-void cutlass_scaled_mm_sm80_dispatch(torch::Tensor& out, torch::Tensor const& a,
-                                     torch::Tensor const& b,
-                                     torch::Tensor const& a_scales,
-                                     torch::Tensor const& b_scales) {
+template <typename InType, typename OutType,
+          template <typename, typename> typename Epilogue,
+          typename... EpilogueArgs>
+void cutlass_gemm_sm80_dispatch(torch::Tensor& out, torch::Tensor const& a,
+                                torch::Tensor const& b,
+                                EpilogueArgs&&... args) {
   static_assert(std::is_same<InType, int8_t>());
   TORCH_CHECK(a.dtype() == torch::kInt8);
   TORCH_CHECK(b.dtype() == torch::kInt8);
-  TORCH_CHECK(a_scales.dtype() == torch::kFloat32);
-  TORCH_CHECK(b_scales.dtype() == torch::kFloat32);
 
   static const int32_t MDimDontCare = 0;
 
   using Cutlass2xGemmDefault =
-      typename sm80_config<InType, OutType, ScaledEpilogue,
+      typename sm80_config<InType, OutType, Epilogue,
                            MDimDontCare>::Cutlass2xGemm;
   using Cutlass2xGemmM128BigN =
-      typename sm80_config<InType, OutType, ScaledEpilogue,
+      typename sm80_config<InType, OutType, Epilogue,
                            MDimDontCare>::Cutlass2xGemm;
   using Cutlass2xGemmM128SmallN =
-      typename sm80_config<InType, OutType, ScaledEpilogue, 64>::Cutlass2xGemm;
+      typename sm80_config<InType, OutType, Epilogue, 64>::Cutlass2xGemm;
   using Cutlass2xGemmM64 =
-      typename sm80_config<InType, OutType, ScaledEpilogue, 64>::Cutlass2xGemm;
+      typename sm80_config<InType, OutType, Epilogue, 64>::Cutlass2xGemm;
   using Cutlass2xGemmM32 =
-      typename sm80_config<InType, OutType, ScaledEpilogue, 32>::Cutlass2xGemm;
+      typename sm80_config<InType, OutType, Epilogue, 32>::Cutlass2xGemm;
   using Cutlass2xGemmM16 =
-      typename sm80_config<InType, OutType, ScaledEpilogue, 16>::Cutlass2xGemm;
+      typename sm80_config<InType, OutType, Epilogue, 16>::Cutlass2xGemm;
 
   uint32_t const m = a.size(0);
   uint32_t const mp2 =
       std::max(static_cast<uint32_t>(16), next_pow_2(m));  // next power of 2
   if (mp2 <= 16) {
-    return cutlass_gemm_caller<Cutlass2xGemmM16>(out, a, b, a_scales, b_scales);
+    return cutlass_gemm_caller<Cutlass2xGemmM16>(
+        out, a, b, std::forward<EpilogueArgs>(args)...);
   } else if (mp2 <= 32) {
-    return cutlass_gemm_caller<Cutlass2xGemmM32>(out, a, b, a_scales, b_scales);
+    return cutlass_gemm_caller<Cutlass2xGemmM32>(
+        out, a, b, std::forward<EpilogueArgs>(args)...);
   } else if (mp2 <= 64) {
-    return cutlass_gemm_caller<Cutlass2xGemmM64>(out, a, b, a_scales, b_scales);
+    return cutlass_gemm_caller<Cutlass2xGemmM64>(
+        out, a, b, std::forward<EpilogueArgs>(args)...);
   } else if (mp2 <= 128) {
     uint32_t const n = a.size(1);
     bool const small_n = n < 8192;
     if (small_n) {
-      return cutlass_gemm_caller<Cutlass2xGemmM128SmallN>(out, a, b, a_scales,
-                                                          b_scales);
+      return cutlass_gemm_caller<Cutlass2xGemmM128SmallN>(
+          out, a, b, std::forward<EpilogueArgs>(args)...);
     } else {
-      return cutlass_gemm_caller<Cutlass2xGemmM128BigN>(out, a, b, a_scales,
-                                                        b_scales);
+      return cutlass_gemm_caller<Cutlass2xGemmM128BigN>(
+          out, a, b, std::forward<EpilogueArgs>(args)...);
     }
   } else {
-    return cutlass_gemm_caller<Cutlass2xGemmDefault>(out, a, b, a_scales,
-                                                     b_scales);
+    return cutlass_gemm_caller<Cutlass2xGemmDefault>(
+        out, a, b, std::forward<EpilogueArgs>(args)...);
   }
 }
 
@@ -390,11 +392,12 @@ void cutlass_scaled_mm_sm80(torch::Tensor& out, torch::Tensor const& a,
   TORCH_CHECK(b_scales.dtype() == torch::kFloat32);
 
   if (out.dtype() == torch::kBFloat16) {
-    return cutlass_scaled_mm_sm80_dispatch<int8_t, cutlass::bfloat16_t>(
-        out, a, b, a_scales, b_scales);
+    return cutlass_gemm_sm80_dispatch<int8_t, cutlass::bfloat16_t,
+                                      ScaledEpilogue>(out, a, b, a_scales,
+                                                      b_scales);
   } else {
     TORCH_CHECK(out.dtype() == torch::kFloat16);
-    return cutlass_scaled_mm_sm80_dispatch<int8_t, cutlass::half_t>(
+    return cutlass_gemm_sm80_dispatch<int8_t, cutlass::half_t, ScaledEpilogue>(
         out, a, b, a_scales, b_scales);
   }
 }