@varun-sundar-rabindranath 2024-06-05
PFA Heatmap generated from the benchmark sweeps on A100.
![model_bench-torch int8-1716411533_heatmap](https://github.com/vllm-project/vllm/assets/3337719/18656059-0ba9-486f-8795-933d4e2ece9e)

Pointers on how to read the heatmap:
X-Axis - All the Cutlass OP tried.
Y-Axis - All the GEMM Shapes in (M x N x K) format.
The Darker the cell, the Better the algorithm performed for that GEMM shape. Each row (GEMM-Shape) is normalized to be between 0.0 and 1.0. The best algorithm has a value of 1.0.

Cutlass Op naming convention:
`autogen_cutlass2x_scaled_mm_dq_sm80_128x64x128_64x64x64_16x8x32_ThreadBlockSwizzleStrreamK_kGemmSplitKParallel_5` refers to an Op constructed with, 
 - Tile Shape : 128x64x128
 - Warp Shape : 64x64x64
 - Instruction Shape : 16x8x32
 - Thread block swizzle : ThreadBlockSwizzleStreamK
 - Gemm mode : kGemmSplitKParallel
 - Main loop stages : 5